Introduce new structure.
A new package structure communicates better what it is about.

The name of the core class now better describes what is actually done.
Also changing from static method to using an instance gains more
flexibility for the future.

Introduce a custom exception which makes it easier for clients to catch
the correct one.